[feat] 소스코드 제출 시 S3 업로드 및 채점 결과 비동기 처리 기능 구현

+++ src/main/java/com/koreii/algoduck/submission/dto/response/JudgeResponseDto.java
@@ -13,6 +13,6 @@ public class JudgeResponseDto {
   private String message;
   private String stdout;
   private String stderr;
-  private double executionTime;
-  private Long memoryUsage;
+  private Integer executionTime;
+  private Integer memoryUsage;
 }
